[#119]: moved read packet functions to dedicated consumer.rs module
